requirements.txt
asgiref==3.5.0
Django==4.0.2
django-widget-tweaks==1.4.12
sqlparse==0.4.2
tzdata==2021.5
whitenoise==6.0.0
django-mptt==0.14.0
django-numerators==0.0.2
django-polymorphic==3.1.0
Pillow==9.3.0
